Finished AStarAI and got Entity Collision working with different size parameters to look better
Added Debug Timer
-Call Timer.startTimer(); before whatever function you want to call
-Call Timer.stopTimer(String of what shows up in console); after to get
the time
-If you call startTimer() 2 times before you call stop, in different
thread, for example, the second call blocks until the first thread calls
stopTimer();

+++ src/net/awesome/game/entities/components/AIComponent.java
@@ -6,7 +6,7 @@ public class AIComponent implements Component{
 	public AI aiModule;
 	public int x, y;
 	public int agroDistance = 10;
-	public int time = 0;
+	public float time = 0;
 	public AIComponent(AI aiModule){
 		this.aiModule = aiModule;
 	}
